+++ spec/www/spec/db-tx-value-bindings-test.js
@@ -531,13 +531,13 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        // TBD ENCODING ISSUE iwth emojis and other 4-octet UTF-8 characters
-        // on Android (default Android evcore NDK implementation)
+        // ENCODING ISSUE NOW FIXED with emojis and other 4-byte UTF-8 characters
+        // on default Android evcore NDK implementation
         // ref:
         // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
         // - litehelpers/Cordova-sqlite-storage#564
         // - litehelpers/Cordova-sqlite-evcore-extbuild-free#7
-        it(suiteName + 'INSERT TEXT string with emoji [\\u1F603 SMILING FACE (MOUTH OPEN)], SELECT the data, check, and check HEX [ENCODING ISSUE REPRODUCED on Android post-5.x; non-standard encoding reproduced on Android pre-6.0; TRUNCATION BUG REPRODUCED on Windows; default sqlite HEX encoding: UTF-6le on Windows & Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]' , function(done) {
+        it(suiteName + 'INSERT TEXT string with emoji [\\u1F603 SMILING FACE (MOUTH OPEN)], SELECT the data, check, and check HEX [non-standard encoding reproduced on Android pre-6.0; TRUNCATION BUG REPRODUCED on Windows; default sqlite HEX encoding: UTF-6le on Windows & Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]' , function(done) {
           var db = openDatabase('INSERT-emoji-and-check.db', '1.0', 'Demo', DEFAULT_SIZE);
 
           db.transaction(function(tx) {
@@ -555,15 +555,9 @@ var mytests = function() {
 
                   var row = rs2.rows.item(0);
                   // Full object check:
-                  if (!isWebSql && !isWindows && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
-                    expect(row).toEqual({data: '@?!'});
-                  else
-                    expect(row).toEqual({data: '@\uD83D\uDE03!'});
+                  expect(row).toEqual({data: '@\uD83D\uDE03!'});
                   // Check individual members:
-                  if (!isWebSql && !isWindows && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
-                    expect(row.data).toBe('@?!');
-                  else
-                    expect(row.data).toBe('@\uD83D\uDE03!');
+                  expect(row.data).toBe('@\uD83D\uDE03!');
 
                   tx.executeSql('SELECT HEX(data) AS hexvalue FROM test_table', [], function(tx_ignored, rs3) {
                     expect(rs3).toBeDefined();
@@ -574,8 +568,6 @@ var mytests = function() {
                       expect(rs3.rows.item(0).hexvalue).toBe('40003DD803DE2100'); // (UTF-16le)
                     else if (!isWebSql && !isWindows && isAndroid && !isImpl2 && /Android [4-5]/.test(navigator.userAgent))
                       expect(rs3.rows.item(0).hexvalue).toBe('40EDA0BDEDB88321'); // non-standard encoding on Android pre-6.0
-                    else if (!isWebSql && isAndroid && !isImpl2)
-                      expect(rs3.rows.item(0).hexvalue).toBe('403F21'); // ENCODING ISSUE on Anroid post-5.x UTF-8
                     else
                       expect(rs3.rows.item(0).hexvalue).toBe('40F09F988321'); // (UTF-8)
 
@@ -598,7 +590,7 @@ var mytests = function() {
         // ref:
         // - litehelpers/Cordova-sqlite-evcore-extbuild-free#43
         // - litehelpers/Cordova-sqlite-storage#564
-        it(suiteName + 'INSERT TEXT string with 25 emojis, SELECT the data, and check - ENCODING ISSUE REPRODUCED on Android post-5.x (default evcore NDK implementation)' , function(done) {
+        it(suiteName + 'INSERT TEXT string with 25 emojis, SELECT the data, and check' , function(done) {
           // ref:
           // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
           // - litehelpers/Cordova-sqlite-storage#564
@@ -621,10 +613,7 @@ var mytests = function() {
                   expect(rs2.rows.length).toBe(1);
 
                   var row = rs2.rows.item(0);
-                  if (!isWebSql && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
-                    expect(row.data).toBe('@?????@?????@?????@?????@?????');
-                  else
-                    expect(row.data).toBe(part + part + part + part + part);
+                  expect(row.data).toBe(part + part + part + part + part);
 
                   // Close (plugin only) & finish:
                   (isWebSql) ? done() : db.close(done, done);
@@ -1165,7 +1154,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + "INSERT inline BLOB value (X'FFD1FFD2') and check stored data [XXX PLUGIN ISSUES REPRODUCED: SELECT BLOB returns data with incorrect length on Android (default evcore NDK implementation); SELECT BLOB VALUE ERROR on Android (androidDatabaseProvider: 'system') & Windows; MISSING result data column on iOS/macOS; actual result value is IGNORED on (WebKit) Web SQL & plugin on other platforms]", function(done) {
+        it(suiteName + "INSERT inline BLOB value (X'FFD1FFD2') and check stored data [PLUGIN ISSUES REPRODUCED: SELECT BLOB returns data with incorrect length on Android (default evcore NDK implementation); SELECT BLOB VALUE ERROR on Android (androidDatabaseProvider: 'system') & Windows; SKIP SELECT BLOB VALUE TEST on default Android evcore NDK implementation TO AVOID KNOWN CRASH; MISSING result data column on iOS/macOS; actual result value is IGNORED on (WebKit) Web SQL & plugin on other platforms]", function(done) {
           var db = openDatabase('INSERT-inline-BLOB-value-FFD1FFD2-and-check-stored-data.db', '1.0', 'Demo', DEFAULT_SIZE);
 
           db.transaction(function(tx) {
@@ -1185,6 +1174,9 @@ var mytests = function() {
                   var item = rs2.rows.item(0);
                   expect(item).toBeDefined();
                   expect(item.hexValue).toBe('FFD1FFD2');
+
+                  // STOP HERE for default Android evcore NDK implementation TO AVOID KNOWN CRASH:
+                  if (!isWebSql && isAndroid && !isImpl2) return done();
 
                   tx.executeSql('SELECT * FROM test_table', [], function(ignored, rs3) {
                     if (!isWebSql && isWindows) expect('PLUGIN BEHAVIOR CHANGED: UNEXPECTED SUCCESS on Windows').toBe('--');