+++ CHANGES.md
@@ -2,6 +2,19 @@
 
 ## cordova-sqlite-evcore-feat-android-db-location 0.9.0-dev
 
+### cordova-sqlite-evcore-common-free 0.5.0
+
+- update evcore NDK library - solution for Android 11 with target SDK 30, from cordova-sqlite-evcore-free-dependencies@0.11.0
+
+### cordova-sqlite-evcore-commoncore-free 0.4.0
+
+- updates from cordova-sqlite-evcore-free-dependencies@0.10.1
+  - BREAKING: drop Android pre-5.1 support
+  - SQLite 3.35.5 update
+  - fix for emojis and other 4-byte UTF-8 characters on Android
+  - update affected test cases
+  - update README.md
+
 ## cordova-sqlite-evcore-feat-android-db-location 0.8.0
 
 - refactor: update androidDatabaseLocation error messages